Add fromProperties() factory constructors
